+++ app/models/post_comment.rb
@@ -1,7 +1,6 @@
 class PostComment < ApplicationRecord
   belongs_to :user
   belongs_to :post
-  has_one_attached :image
 
   validates :content, presence: true
 end
