+++ app/views/public/users/edit.html.erb
@@ -26,8 +26,13 @@
     <% end %>
   </div>
 
+  <div class="field">
+    <%= f.label :self_introduction %><br>
+    <%= f.text_area :self_introduction %>
+  </div>
+
   <div class="actions">
-    <%= f.submit "更新" %>
+    <%= f.submit "更新する" %>
   </div>
 <% end %>
