+++ config/routes.rb
@@ -14,6 +14,7 @@ Rails.application.routes.draw do
   # 投稿関連
   scope module: :public do
     resources :posts do
+      resources :post_comments, only: [:create, :destroy]
       collection do
         get 'search'
       end
